+++ src/main/java/com/project/Teaming/domain/mentoring/service/MentoringParticipationService.java
@@ -1,6 +1,5 @@
 package com.project.Teaming.domain.mentoring.service;
 
-import com.project.Teaming.domain.mentoring.annotation.NotifyAfterTransaction;
 import com.project.Teaming.domain.mentoring.dto.request.ParticipationRequest;
 import com.project.Teaming.domain.mentoring.dto.response.*;
 import com.project.Teaming.domain.mentoring.entity.*;
@@ -13,6 +12,7 @@ import com.project.Teaming.domain.user.entity.User;
 import com.project.Teaming.domain.user.repository.ReportRepository;
 import com.project.Teaming.domain.user.service.ReportService;
 import com.project.Teaming.domain.user.service.ReviewService;
+import com.project.Teaming.global.annotation.NotifyAfterTransaction;
 import com.project.Teaming.global.error.ErrorCode;
 import com.project.Teaming.global.error.exception.*;
 import com.project.Teaming.global.sse.service.NotificationService;
